splitting out form demo page into form and dynamic form
