+++ projects/novo-examples/src/form-controls/ck-editor/basic-editor/basic-editor-example.css
