+++ projects/novo-examples/src/form-controls/ck-editor/basic-editor/index.ts
