+++ projects/novo-examples/src/form-controls/ck-editor/minimal-editor/index.ts
