+++ projects/novo-examples/src/form-controls/ck-editor/minimal-editor/minimal-editor-example.ts
