+++ projects/novo-examples/src/form-controls/dynamic-form/dynamic-form.md
@@ -1,0 +1,34 @@
+Dynamic Forms [(source)](https://github.com/bullhorn/novo-elements/blob/master/projects/novo-elements/src/elements/form)
+=========================================================================================
+
+Forms use inputs and labels to submit user content. But you already knew that. What you may not know is that our forms come in two styles 'Static' and 'Dynamic'.
+This is the Dynamic Form demo.
+
+Dynamic Form
+------------
+
+Dynamic forms are composed of one element, `<novo-dynamic-form [controls]="controls"/>` and allow you to pass in the controls and form and it will create the form for you.
+
+##### Basic
+
+<code-example example="dynamic-form"></code-example>
+
+##### Vertical
+
+<code-example example="vertical-dynamic-form"></code-example>
+
+##### Fieldsets
+
+<code-example example="dynamic-form-field-sets"></code-example>
+
+##### Updating Fields/Status
+
+<code-example example="updating-form"></code-example>
+
+##### Disabled Field States
+
+<code-example example="disabled-form"></code-example>
+
+##### Enable/Disable All Fields in Form
+
+<code-example example="enable-disable-all-fields-in-form"></code-example>
