+++ projects/novo-examples/src/form-controls/form/check-box-controls/check-box-controls-example.ts
@@ -1,8 +1,5 @@
 import { Component } from '@angular/core';
-// Vendor
 import { CheckboxControl, CheckListControl, FormUtils, SwitchControl, TilesControl } from 'novo-elements';
-
-// import { MockMeta, MockMetaHeaders } from './MockMeta';
 
 /**
  * @title Check Box Controls Example
