+++ projects/novo-examples/src/form-controls/form/index.ts
@@ -2,12 +2,6 @@ export * from './address-control';
 export * from './calendar-input-controls';
 export * from './check-box-controls';
 export * from './custom-demo-control';
-export * from './disabled-form';
-export * from './dynamic-form';
-export * from './dynamic-form-field-sets';
-export * from './enable-disable-all-fields-in-form';
 export * from './file-input-controls';
 export * from './picker-controls';
 export * from './text-based-controls';
-export * from './updating-form';
-export * from './vertical-dynamic-form';
