+++ projects/novo-examples/src/form-controls/index.ts
@@ -1,8 +1,9 @@
 export * from './checkbox';
 export * from './chips';
+export * from './ck-editor';
 export * from './date-picker';
 export * from './date-time-picker';
-export * from './editor';
+export * from './dynamic-form';
 export * from './form';
 export * from './form-groups';
 export * from './multi-picker';
